backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers
djangorestframework-simplejwt
transformers==4.36.2
torch
